+++ R/cj_anova.R
@@ -19,6 +19,7 @@ function(
     by_vars <- all.vars(stats::update(by, 0 ~ . ))
     
     # modify formula to include appropriate interaction
+    formula_reduced <- formula
     formula_full <- update(formula, reformulate(paste0("(.) * ", by_vars)))
     
     # get `id` as character string
@@ -26,17 +27,19 @@ function(
     
     # estimate model
     if (inherits(data, "data.frame") && is.null(weights)) {
-        svydesign <- NULL
+        svy_design <- NULL
         estimates_full <- stats::glm(formula_full, data = data, ...)
-        estimates_reduced <- stats::glm(formula, data = data, ...)
+        estimates_reduced <- stats::glm(formula_reduced, data = data, ...)
     } else if (inherits(data, "data.frame")) {
-        svydesign <- survey::svydesign(ids = ~ 0, weights = weights, data = data)
-        estimates_full <- survey::svyglm(formula_full, design = svydesign, ...)
-        estimates_reduced <- survey::svyglm(formula, design = svydesign, ...)
+        stop("cj_anova() currently does not support 'weights'")
+        svy_design <- survey::svydesign(ids = ~ 0, weights = weights, data = data)
+        estimates_full <- survey::svyglm(formula_full, design = svy_design, ...)
+        estimates_reduced <- survey::svyglm(formula_reduced, design = svy_design, ...)
     } else if (inherits(data, "survey.design")) {
-        svydesign <- data
-        estimates_full <- survey::svyglm(formula_full, design = svydesign, ...)
-        estimates_reduced <- survey::svyglm(formula, design = svydesign, ...)
+        stop("cj_anova() currently does not support 'weights'")
+        svy_design <- data
+        estimates_full <- survey::svyglm(formula_full, design = svy_design, ...)
+        estimates_reduced <- survey::svyglm(formula_reduced, design = svy_design, ...)
     } else {
         stop("'data' is not a 'data.frame' or 'survey.design' object")
     }